fix: remove redundant autoload and require
